+++ asteroid-test/src/main/groovy/asteroid/global/samples/ChangeMethodsTransformation.groovy
@@ -2,8 +2,7 @@ package asteroid.global.samples
 
 import groovy.transform.CompileStatic
 
-import asteroid.A
-import asteroid.GlobalTransformation
+import asteroid.Phase
 import asteroid.AbstractGlobalTransformation
 import asteroid.transformer.Transformer
 
@@ -14,7 +13,7 @@ import asteroid.transformer.Transformer
  * @since 0.1.2
  */
 @CompileStatic
-@GlobalTransformation(A.PHASE_GLOBAL.CANONICALIZATION)
+@Phase(Phase.GLOBAL.CANONICALIZATION)
 class ChangeMethodsTransformation extends AbstractGlobalTransformation {
 
     @Override