+++ asteroid-test/src/main/groovy/asteroid/local/samples/AddElementsImpl.groovy
@@ -1,7 +1,7 @@
 package asteroid.local.samples
 
 import asteroid.A
-import asteroid.LocalTransformation
+import asteroid.Phase
 import asteroid.AbstractLocalTransformation
 
 import groovy.transform.CompileStatic
@@ -12,7 +12,7 @@ import org.codehaus.groovy.ast.PropertyNode
 import org.codehaus.groovy.ast.expr.Expression
 
 @CompileStatic
-@LocalTransformation(A.PHASE_LOCAL.SEMANTIC_ANALYSIS)
+@Phase(Phase.LOCAL.SEMANTIC_ANALYSIS)
 class AddElementsImpl extends AbstractLocalTransformation<AsList, ClassNode> {
 
     @Override