+++ asteroid-test/src/main/groovy/asteroid/local/samples/ExecuteMethodSafelyImpl.groovy
@@ -1,7 +1,7 @@
 package asteroid.local.samples
 
 import asteroid.A
-import asteroid.LocalTransformation
+import asteroid.Phase
 import asteroid.AbstractLocalTransformation
 
 import groovy.transform.CompileStatic
@@ -10,7 +10,7 @@ import org.codehaus.groovy.ast.AnnotationNode
 import org.codehaus.groovy.ast.FieldNode
 
 @CompileStatic
-@LocalTransformation(A.PHASE_LOCAL.SEMANTIC_ANALYSIS)
+@Phase(Phase.LOCAL.SEMANTIC_ANALYSIS)
 class ExecuteMethodSafelyImpl extends AbstractLocalTransformation<ExecuteMethodSafely, MethodNode> {
 
     @Override