+++ asteroid-test/src/main/groovy/asteroid/local/samples/ToMD5Impl.groovy
@@ -1,7 +1,7 @@
 package asteroid.local.samples
 
 import asteroid.A
-import asteroid.LocalTransformation
+import asteroid.Phase
 import asteroid.AbstractLocalTransformation
 
 import groovy.transform.CompileStatic
@@ -12,7 +12,7 @@ import org.codehaus.groovy.ast.MethodNode
 import org.codehaus.groovy.ast.stmt.BlockStatement
 
 @CompileStatic
-@LocalTransformation(A.PHASE_LOCAL.SEMANTIC_ANALYSIS) // <1>
+@Phase(Phase.LOCAL.SEMANTIC_ANALYSIS) // <1>
 class ToMD5Impl extends AbstractLocalTransformation<ToMD5, FieldNode> { // <2>
 
     @Override