+++ asteroid-test/src/main/groovy/asteroid/local/samples/ToMapImpl.groovy
@@ -1,7 +1,7 @@
 package asteroid.local.samples
 
 import asteroid.A
-import asteroid.LocalTransformation
+import asteroid.Phase
 import asteroid.AbstractLocalTransformation
 
 import groovy.transform.CompileStatic
@@ -14,7 +14,7 @@ import org.codehaus.groovy.ast.expr.MapExpression
 import org.codehaus.groovy.ast.expr.MapEntryExpression
 
 @CompileStatic
-@LocalTransformation(A.PHASE_LOCAL.INSTRUCTION_SELECTION)
+@Phase(Phase.LOCAL.INSTRUCTION_SELECTION)
 class ToMapImpl extends AbstractLocalTransformation<ToMap, ClassNode> {
 
     @Override