+++ asteroid-test/src/main/groovy/asteroid/local/samples/WithLogging.groovy
@@ -1,10 +1,7 @@
 package asteroid.local.samples
 
 import static asteroid.A.TO
-
 import asteroid.Local
-import asteroid.Apply
 
-@Apply(TO.METHOD) // <1>
-@Local(WithLoggingTransformationImpl) // <2>
+@Local(value = WithLoggingTransformationImpl, applyTo = TO.METHOD) // <2>
 @interface WithLogging { }